requirements.txt
python-dotenv
fastapi[standard-no-fastapi-cloud-cli]
slack-bolt
aiohttp
pydantic-ai-slim[openai,anthropic,logfire,evals,duckduckgo,tavily]
logfire[fastapi,httpx]
twilio
pyngrok
